Extract TCP Operations to their own header
