+++ scaler/io/ymq/io_socket.h
@@ -15,6 +15,7 @@
 
 class EventLoopThread;
 class MessageConnectionTCP;
+class TcpWriteOperation;
 
 class IOSocket {
 public:
